Bug 574848 Extract plain Passage implementation out of PassageEquinox
BasePassage is invited built over framework retrieved by client
