+++ bundles/org.eclipse.passage.lic.base/src/org/eclipse/passage/lic/base/BasePassage.java
@@ -10,54 +10,49 @@
  * Contributors:
  *     ArSysOp - initial API and implementation
  *******************************************************************************/
-package org.eclipse.passage.lic.equinox;
+package org.eclipse.passage.lic.base;
 
 import org.eclipse.passage.lic.api.LicensedProduct;
 import org.eclipse.passage.lic.api.Passage;
 import org.eclipse.passage.lic.api.ServiceInvocationResult;
 import org.eclipse.passage.lic.api.access.GrantLockAttempt;
 import org.eclipse.passage.lic.api.restrictions.ExaminationCertificate;
-import org.eclipse.passage.lic.base.BasePassage;
-import org.eclipse.passage.lic.base.FrameworkAware;
+import org.eclipse.passage.lic.base.access.Access;
 
 /**
- * @since 2.1
+ * @since 1.1
  */
-public final class EquinoxPassage implements Passage {
+public final class BasePassage implements Passage {
 
-	private final BasePassage delegate;
+	private final FrameworkAware delegate;
 
-	public EquinoxPassage() {
-		this(new SuppliedFrameworkAware());
-	}
-
-	public EquinoxPassage(FrameworkAware delegate) {
-		this.delegate = new BasePassage(delegate);
+	public BasePassage(FrameworkAware delegate) {
+		this.delegate = delegate;
 	}
 
 	@Override
 	public boolean canUse(String feature) {
-		return delegate.canUse(feature);
+		return delegate.withFramework(framework -> new Access(framework).canUse(feature)).orElse(Boolean.FALSE);
 	}
 
 	@Override
 	public ServiceInvocationResult<ExaminationCertificate> assess() {
-		return delegate.assess();
+		return delegate.withFrameworkService(framework -> new Access(framework).assess());
 	}
 
 	@Override
 	public ServiceInvocationResult<GrantLockAttempt> acquireLicense(String feature) {
-		return delegate.acquireLicense(feature);
+		return delegate.withFrameworkService(framework -> new Access(framework).acquire(feature));
 	}
 
 	@Override
 	public ServiceInvocationResult<Boolean> releaseLicense(GrantLockAttempt lock) {
-		return delegate.releaseLicense(lock);
+		return delegate.withFrameworkService(framework -> new Access(framework).release(lock));
 	}
 
 	@Override
 	public ServiceInvocationResult<LicensedProduct> product() {
-		return delegate.product();
+		return delegate.withFrameworkService(framework -> new BaseServiceInvocationResult<>(framework.product()));
 	}
 
 }
